Try to fix destroy again

+++ plugins/orders/controllers/myprofile/orders_plugin_item_controller.rb
@@ -36,6 +36,7 @@ class OrdersPluginItemController < MyProfileController
 
   def destroy
     item = OrdersPlugin::Item.where(id: params[:id]).first
+    return unless item
     @offered_product = item.offered_product
     item.destroy
   end
